add functionality to set a pre-calculated marcid
With that hack it is possible to create a record with a pre-calculated marcid.
This approach is necessary because there exists for now some DOI's with marcid (with no
corresponding record in the repository) where the records should
have then also the same marcid when they will be created then in the repository.

To set the pre-calculated marcid it has to be assigned to
MarcDraftProvider.predefined_pid_value before the creation of the record with
the corresponding service