make the jsonschema less restrictive
it could be that there are fields comming from alma which whill have
further subfields or also further indicators that are not in this
jsonschema. the best option would be to add every further field,
subfield, indicator etc if a problem arrise, but that would take a lot
of time. this solution is a more pragmatic solution.